requirements.txt
fastapi
uvicorn
prometheus_client
python-dotenv
redis[asyncio]
pytest
pytest-cov